CDRJTW REFACTOR (cross-scenario): extract shared boundary test fixtures
git()/readAudit()/ticketContent() were triplicated across the four
boundary suites — extracted to tests/commands/boundary-helpers.ts.
Behavior-preserving: 27/27 green, lint clean. (The gate itself flagged
the open cross-scenario row on the previous commit — this is the pass
it asked for.)
