+++ packages/cli/tests/commands/boundary-push.test.ts
@@ -5,8 +5,6 @@
  */
 
 import { execSync } from 'node:child_process';
-import { existsSync, readFileSync } from 'node:fs';
-import nodePath from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
@@ -17,42 +15,9 @@ import {
   runCli,
   writeTestFile,
 } from '../helpers';
+import { boundaryTicketContent as ticketContent, git, readAudit } from './boundary-helpers';
 
-const AUDIT_PATH = '.safeword/boundary-audit.jsonl';
 const TICKET = '.project/tickets/BNP001-fixture';
-
-function git(dir: string, command: string): string {
-  return execSync(`git ${command}`, { cwd: dir, stdio: 'pipe', encoding: 'utf8' });
-}
-
-function readAudit(dir: string): Record<string, unknown>[] {
-  const auditFile = nodePath.join(dir, AUDIT_PATH);
-  if (!existsSync(auditFile)) return [];
-  return readFileSync(auditFile, 'utf8')
-    .split('\n')
-    .filter(line => line.trim() !== '')
-    .map(line => JSON.parse(line) as Record<string, unknown>);
-}
-
-function ticketContent(options: { phase: string; anchors?: string[]; skips?: string[] }): string {
-  const lines = [
-    '---',
-    'id: ZZBNP',
-    'type: feature',
-    `phase: ${options.phase}`,
-    'status: in_progress',
-  ];
-  if (options.anchors) {
-    lines.push('phase_anchors:');
-    for (const entry of options.anchors) lines.push(`  - ${entry}`);
-  }
-  if (options.skips) {
-    lines.push('phase_skips:');
-    for (const entry of options.skips) lines.push(`  - ${entry}`);
-  }
-  lines.push('---', '', '# Fixture', '');
-  return lines.join('\n');
-}
 
 describe('safeword boundary (slice 3: push tier)', () => {
   let dir: string;
